General cleanup; upgraded prime test to use Long instead of Int

+++ src/main/java/org/hackreduce/examples/prime/PrimeTest.java
@@ -1,15 +1,10 @@
 package org.hackreduce.examples.prime;
 
 import java.io.IOException;
-import java.text.NumberFormat;
-import java.util.Locale;
-
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.DoubleWritable;
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
@@ -31,9 +26,6 @@ public class PrimeTest extends Configured implements Tool {
 
 		@Override
 		protected void map(PrimeRecord record, Context context) throws IOException, InterruptedException {
-			//if (record.getNum()%2 == 0) {
-				//context.write(new Text(Integer.toString(record.getNum())), new LongWritable(1));
-			//}
 			boolean flag = true;
 			for(int i=2; i<=Math.sqrt(record.getNum()); i++) {
 	              if (record.getNum()%i == 0) {
@@ -42,7 +34,7 @@ public class PrimeTest extends Configured implements Tool {
 	              }
 	        }
 			if (flag) {
-				context.write(new Text(Integer.toString(record.getNum())), new LongWritable(1));
+				context.write(new Text(Long.toString(record.getNum())), new LongWritable(1));
 			} else {
 				// Do nothing.  Number is not prime.  Boo
 			}
@@ -79,7 +71,7 @@ public class PrimeTest extends Configured implements Tool {
         job.setMapperClass(PrimeTestMapper.class);
 		job.setReducerClass(PrimeTestReducer.class);
 
-		// Configure the job to accept the NASDAQ/NYSE dividend data as input
+		// Configure the job to accept the number list data as input
 		PrimeMapper.configureJob(job);
 
 		// This is what the Mapper will be outputting to the Reducer
